add Abstract class, implement inheritance for all components, add page controller for cards' rendering
