+++ src/components/no-films.js
@@ -1,22 +1,9 @@
-import {createElement} from "../utils";
+import {AbstractComponent} from "./absctract-component";
 
-export class NoFilms {
+export class NoFilms extends AbstractComponent {
   constructor() {
-    this._element = null;
+    super();
   }
-
-  getElement() {
-    if (!this._element) {
-      this._element = createElement(this.getTemplate());
-    }
-
-    return this._element;
-  }
-
-  removeElement() {
-    this._element = null;
-  }
-
   getTemplate() {
     return `
       <section class="films">
