+++ src/components/search.js
@@ -1,18 +1,8 @@
-import {createElement} from "../utils";
+import {AbstractComponent} from "./absctract-component";
 
-export class SearchBar {
+export class SearchBar extends AbstractComponent {
   constructor() {
-    this._element = null;
-  }
-  getElement() {
-    if (!this._element) {
-      this._element = createElement(this.getTemplate());
-    }
-
-    return this._element;
-  }
-  removeElement() {
-    this._element = null;
+    super();
   }
   getTemplate() {
     return `
